+++ src/test/cli-options.test.ts
@@ -11,7 +11,6 @@ import {timeout} from './util/uvu-timeout.js';
 import {WireitTestRig} from './util/test-rig.js';
 import {Options} from '../cli-options.js';
 import {Result} from '../error.js';
-import {Failure} from '../event.js';
 
 const test = suite<{rig: WireitTestRig}>();
 
@@ -84,19 +83,6 @@ async function assertOptions(
       failureMode: 'no-new',
       ...expected,
     },
-  });
-}
-
-async function assertFailure(
-  rig: WireitTestRig,
-  command: string,
-  expected: Failure,
-  env?: Record<string, string | undefined>
-) {
-  const result = await getOptionsResult(rig, command, env);
-  assert.equal(result, {
-    ok: false,
-    error: expected,
   });
 }
 
@@ -274,24 +260,6 @@ for (const command of ['npm', 'yarn', 'pnpm'] as const) {
         },
         extraArgs: ['--extra'],
         watch: true,
-      });
-    })
-  );
-
-  test(
-    `temporary error on ${command} run main watch`,
-    timeout(async ({rig}) => {
-      await assertFailure(rig, `${command} run main watch`, {
-        script: {
-          packageDir: rig.temp,
-          name: 'main',
-        },
-        type: 'failure',
-        reason: 'invalid-usage',
-        message:
-          `As of wireit v0.6, use "--watch" instead of "watch". ` +
-          `In an upcoming release, the "watch" argument will be passed to the script, ` +
-          `consistent with how npm usually behaves.`,
       });
     })
   );